移植到 TC 7.15 (Dalamud API12)
從 PunishXIV/Questionable 的 LLib→ECommons 重構版本(2026-05-04)為基礎,
把 Dalamud.NET.Sdk 從 15.0.0 降到 12.0.2,配合處理三個世代的 API 落差:

- 移除 WrathCombo.API 依賴(該套件從發布以來就只built給 net10,沒有相容
  舊版可用),連帶拿掉 WrathCombo 戰鬥模組,保留 BossMod/RotationSolverReborn
  兩個替代選項
- ImGuiNET 綁定世代(Dalamud.Bindings.ImGui → ImGuiNET)、ITargetManager/
  ISigScanner 命名空間搬遷、IPlayerState 服務移除(改用原生 PlayerState.
  Instance())
- Lumina IExcelRow<T> 介面舊版沒有 ExcelPage/RowOffset 成員
- ImRaii 各種具名 Disposable 型別(TabItemDisposable/ChildDisposable等)
  舊版統一回傳 IEndObject
- 一批 FFXIVClientStructs 欄位/enum 改名(ObjectKind.MountType、
  BattleNpcSubKind.BattleNpcPart、PlayerState.FreeAetherytePlayStationPlus
  等)
- vnavmesh 同款 CameraEx pre-api13 offset(DirH/DirV/InputDeltaH/V)
- ECommons AddonMaster.Shop 這個版本還沒有,改成直接讀 AtkValues 位移
  (照抄 ECommons 自己的實作)
- 一個功能性刪減:AbandonQuest 的 GameMain.ExecuteCommand(900, ...) 在這個
  API 世代找不到對應函式,先停用並提示改用遊戲內日誌手動放棄任務

已用 .NET reflection 確認編譯出來的 DLL 真的連到 Dalamud 12.0.2,不是
不小心連到別的版本。tc-api13-draft 之後再另外處理。

+++ Questionable/Controller/Utils/PartyWatchDog.cs
@@ -31,7 +31,7 @@ internal sealed class PartyWatchDog : IDisposable
         _clientState.TerritoryChanged -= TerritoryChanged;
     }
 
-    private unsafe void TerritoryChanged(uint newTerritoryId)
+    private unsafe void TerritoryChanged(ushort newTerritoryId)
     {
         TerritoryIntendedUseEnum intendedUse = (TerritoryIntendedUseEnum)GameMain.Instance()->CurrentTerritoryIntendedUseId;
         switch (intendedUse)
